Finally!
Renamed to Sherwood, which bears a lexical similarity to the brand of curry
powder in my pantry. (Get it? Curry? Hahaha! Oh, I slay myself...)

Other name considered was Lister, after a certain curry-loving spacefarer.

Chose `.sher` as the file extension -- it seems that it's not being used at all.

Started on standard builtins in `standard.sher`.